updated constructor, read and write methods
changed the argument passed to the constructor to a partridge feed
object and added basic error checking and initialization of class
variables. Updated arguments and comments of read method. Added code
that loads a feed and creates a transit network object. Updated write
method to only write in network standard and not cube